Better structure private items
